completed functionality of deep page extraction

+++ src/main/java/com/domain/searchengine/model/IndexResult.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class IndexResult {
 
-    private int pageAdded;
-    private int keywordsAdded;
+    private long pageAdded;
+    private long keywordsAdded;
 
 }
